Minor Refactoring to clean unused classes
